// src/lib/schema.ts
export const ContactSchema = z.object({
  first_name: z.string().nonempty({ message: "لا يمكن اهمال هذا المدخل" }),
  last_name: z.string().nonempty({ message: "لا يمكن اهمال هذا المدخل" }),
  email: z
    .string()
    .nonempty({ message: "لا يمكن اهمال هذا المدخل" })
    .email({ message: "البريد الالكتروني غير صحيح" }),
  message: z.string().nonempty({ message: "لا يمكن اهمال هذا المدخل" }),
  subject: z.string().nonempty({ message: "لا يمكن اهمال هذا المدخل" })
});